+++ ChangeLog.md
@@ -1,3 +1,6 @@
+# Pending
+* Spot and Hidden effects can have different sources
+
 # v2.2.0
 * Added Brazilian Portuguese support (thanks lucaspicerni)
 * Verify that CUB/CE still has the desired effect before trying to create one with them (#63). Falls back to default in these cases.